Remove no longer available unused safetycounter.h include
kdevplatform/language/duchain/safetycounter.h was recently removed in
https://commits.kde.org/kdevelop/kdevelop/b4eeba21c1d88903fca84642153ae087e129b4a0

The include of the removed header in this plugin was missed. Simply
remove it now, because neither the single declaration in the removed
header `struct SafetyCounter` nor the single include <QDebug> is used
in the affected source file kdev-xdebug/stringhelpers.cpp.

BUG: 467452